Add RawMLS wrapper
This makes it easier to verify signatures and to forward raw serialised
MLS objects. See haddocks for more details.
