+++ libs/wire-api/src/Wire/API/MLS/Proposal.hs
@@ -22,6 +22,7 @@ import Data.Binary.Get
 import Imports
 import Wire.API.Arbitrary
 import Wire.API.MLS.CipherSuite
+import Wire.API.MLS.Extension
 import Wire.API.MLS.Group
 import Wire.API.MLS.KeyPackage
 import Wire.API.MLS.Serialisation
